request_instance: return entire `dict` payload
With the addition of reservations support in this function of the
Python client library, the variety of different responses from the
rerobots API motivates avoiding response-dependent return tuples, and
instead, providing the user a `dict` in the response (i.e., the
payload).

If the API response is expanded, then in future work the `dict` object
returned by request_instance() might be a subset of the response body,
containing only critical values. Also in future work, consider to
provide an option to include diagnostics and debugging details about
the connection etc.