TEST: CI: do not perform static analysis of tests
There are too many false positives. It should be possible to configure
pylint to perform better when checking the code of tests, but it is
not clearly worth the work yet.

In this changeset, static analysis of the code of tests is moved to
the Make target `checktests`, which is not run during CI jobs.