create APIClient methods for `cam` add-on
Each corresponds to an HTTP API command.

status_addon_cam() returns {'status': 'notfound'} if the response has
code 404. This is motivated by the common case of simultaneously
checking whether the add-on applies and, if so, getting its status.
Thus, it is not necessary for the caller to place the call in an
exception-catching block.