create Instance class to manage each instance
This includes methods to setup and execute commands through an SSH
client, as provided by `paramiko`. Methods of APIClient that involve a
particular instance or workspace deployment are wrapped by passing the
corresponding identifier of the Instance object.

Regarding implementation, Instance derives from APIClient. However,
this can cause confusion because some methods of APIClient involve the
user account or other instances. Thus, consider refactoring to

1. have an APIClient attribute of Instance, which is invoked by the
   implementation but also available to the user to invoke;

2. allow user to provide APIClient when creating an Instance object,
   e.g., so multiple Instance objects can be created with the same
   APIClient object.

no tests for it yet, so
[ci skip]