TEST: ignore environ var when not part of testing
Motivating use-case:

User has REROBOTS_API_TOKEN assigned in her current shell. Then, she
`make check`, which is presented in README.rst as the basic command
for testing, but she does not want her personal API token to be
applied for tests that involve rerobots in production. In particular,
this might involve costs to be billed to the user's rerobots account.