Instance can be created by attaching to existing
There are several methods to create a new workspace instance. It can
be useful to create a new Instance object by attaching to an existing
workspace instance, e.g., that was created through the Web UI
(https://rerobots.net/).

This changeset implements the feature, including consistency checks if
the deployment identifier is given.