lint
regarding `elif` after `break`, despite not being permitted by a
pylint rule, there are cases where it is motivated. for example, if
code in the conditional branch before the `elif` is under development,
then it is better to keep the `elif` if it represents the algorithm
accurately (i.e., if the second condition should not be checked unless
the first is false) because the preceding `break` might be removed
during development.