TEST: ignore bad-option-value of pylint
original motivation is to treat case of old versions of pylint failing
on new options in the source code. the actual example is
import-outside-toplevel, which is demonstrated by comparing
https://travis-ci.org/rerobots/py/jobs/597505730
https://travis-ci.org/rerobots/py/jobs/597505732

for this repository, pylint options saved in situ (via comment markup)
are only used to make pylint less constrained. therefore, ignoring
bad-option-value will not result in false negatives.