include capabilities (a.k.a. rules) in `wdinfo`
If given an API token, return capabilities (a.k.a. rules) associated
with that token for the wdeployment given in the CLI command `wdinfo`.
E.g.,

    $ rerobots wdinfo 228238e2-efaf-462d-a5d3-cefc0a2dfbdd

    {
      "id": "228238e2-efaf-462d-a5d3-cefc0a2dfbdd",
      "type": "user_provided",
      "created": "2020-06-20 19:20:24.437513",
      "queuelen": 0,
      "cap": [
	{
	  "id": 15,
	  "date_created": "2020-08-28 07:56:03.730349",
	  "user": "*",
	  "wdeployment_id": "228238e2-efaf-462d-a5d3-cefc0a2dfbdd",
	  "capability": "CAP_INSTANTIATE",

shows that the API token has permission to instantiate the wdeployment
228238e2-efaf-462d-a5d3-cefc0a2dfbdd. This is a special example
because it is of type `user_provided` (i.e., a device available via
hardshare) and has the default of not granting permissions unless a
capability is explicitly granted.